mg: Ensure canvas paint on first run
On first run the main component loads the weather data which triggers
meteogram changed event. The Meteogram component isn't initialized until
the widget is first expanded. In other words it misses the meteogram
changed event so the data isn't processed.

Do a full processing of the data and repaint call on component complete
to ensure the meteogram is all set up properly.