Better cursor tracking
Map the time position to an element in the data model by using search
instead of assuming it is a one to one index.

This resolves the issue with daylight saving where data sources provide
the data for the hour that is skipped. QML time is based on linear time
since epoch. Wall clock time is non-linear when daylight savings
adds/subtracts an hour.